Reference

Contents

Reference

Contents

Index

Index

UCIWWEIHR.NegativeBinomial2

UCIWWEIHR.eihr_ode!

UCIWWEIHR.generate_simulation_data_ww_eihr

UCIWWEIHR.sum_values

UCIWWEIHR.NegativeBinomial2 — Method

Create a re-parametrized negative binomial distribution in terms of mean and overdispersion.

Arguments

μ: Mean of the distribution.

ϕ: Overdispersion parameter.

Returns

A Distributions.NegativeBinomial distribution object.

UCIWWEIHR.eihr_ode! — Method

eihr_ode!(du, u, p, t)

Calculate the ordinary differential equations (ODEs) for the EIHR model.

Parameters:

du: Array{Float64,1} - The derivative of the state variables.

u: Array{Float64,1} - The current state variables.

p: Tuple{Float64,Float64,Float64,Float64,Float64} - The model parameters (alpha, gamma, nu, w, epsilon).

t: Float64 - The current time.

Returns:

du: Array{Float64,1} - The derivative of the state variables.

UCIWWEIHR.generate_simulation_data_ww_eihr — Function

Generating Simulation Data for Wastewater EIHR

To generate simulation data for wastewater EIHR, you can use the generate_simulation_data_ww_eihr function defined in the UCIWWEIHR.jl package. This function allows you to customize various parameters for the simulation.

Function Signature

```julia

Arguments

time_points::Int64: Number of time points wanted for simulation. Default value is 150.

seed::Int64: Seed for random number generation. Default value is 1.

E_init::Int64: Initial number of exposed individuals. Default value is 200.

I_init::Int64: Initial number of infected individuals. Default value is 100.

H_init::Int64: Initial number of hospitalized individuals. Default value is 20.

gamma::Float64: Rate of incubation. Default value is 1/4.

nu::Float64: Rate of leaving the infected compartment. Default value is 1/7.

epsilon::Float64: Rate of hospitalization recovery. Default value is 1/5.

rho_gene::Float64: Contribution of infected individual's pathogen genome into wastewater. Default value is 0.011.

tau::Float64: Scale/variation of the log concentration of pathogen genome in wastewater. Default value is 0.1.

df::Float64: Degrees of freedom for generalized t distribution for log concentration of pathogen genome in wastewater. Default value is 29.

sigma_hosp::Float64: Standard deviation for the negative binomial distribution for hospital data. Default value is 800.

Rt_init::Float64: Initial value of the time-varying reproduction number. Default value is 1.

sigma_Rt::Float64: Standard deviation for random walk of time-varying reproduction number. Default value is sqrt(0.02).

w_init::Float64: Initial value of the time-varying hospitalization rate. Default value is 0.35.

sigma_w::Float64: Standard deviation for random walk of time-varying hospitalization rate. Default value is sqrt(0.02).

UCIWWEIHR.sum_values — Method

sum_values(x,y)

This is an example of Docstring. This function receives two numbers x and y and returns the sum of the squares.

\[x^1 + y^1 \]