Reference

Contents

Reference

Contents

Index

Index

UCIWWEIHR.ChainsCustomIndex

UCIWWEIHR.NegativeBinomial2

UCIWWEIHR.calculate_quantiles

UCIWWEIHR.eihr_ode!

UCIWWEIHR.generate_colors

UCIWWEIHR.generate_simulation_data_agent

UCIWWEIHR.generate_simulation_data_uciwweihr

UCIWWEIHR.power

UCIWWEIHR.save_plots_to_docs

UCIWWEIHR.startswith_any

UCIWWEIHR.uciwweihr_fit

UCIWWEIHR.uciwweihr_model

UCIWWEIHR.uciwweihr_visualizer

UCIWWEIHR.ChainsCustomIndex — Method

ChainsCustomIndex(c::Chains, indices_to_keep::BitMatrix)

Reduce Chains object to only wanted indices.

Function created by Damon Bayer.

UCIWWEIHR.NegativeBinomial2 — Method

Create a re-parametrized negative binomial distribution in terms of mean and overdispersion.

Arguments

μ: Mean of the distribution.

ϕ: Overdispersion parameter.

Returns

A Distributions.NegativeBinomial distribution object.

UCIWWEIHR.calculate_quantiles — Method

calculate_quantiles(df, chain, var_prefix, quantiles)

Calculate quantiles for a given chain and variable prefix. Quantiles can be any user desired quantile.

Function created by Christian Bernal Zelaya.

UCIWWEIHR.eihr_ode! — Method

eihr_ode!(du, u, p, t)

Calculate the ordinary differential equations (ODEs) for the EIHR model.

Parameters:

du: Array{Float64,1} - The derivative of the state variables.

u: Array{Float64,1} - The current state variables.

p: Tuple{Float64,Float64,Float64,Float64,Float64} - The model parameters (alpha, gamma, nu, w, epsilon).

t: Float64 - The current time.

Returns:

du: Array{Float64,1} - The derivative of the state variables.

UCIWWEIHR.generate_colors — Method

generate_ribbon_colors(number_of_colors)

Generates a vector with colors for ribbons in plots.

Function created by Christian Bernal Zelaya.

UCIWWEIHR.generate_simulation_data_agent — Function

Generating Simulation Data for Agent Based Model

To generate simulation data using the agent based model, you can use the generate_simulation_data_agent function defined in the UCIWWEIHR.jl package. This function allows you to customize various parameters for the simulation. NOT FINISHED, STILL NEEDS WW AND RT

Function Signature

Arguments

seed::Int64: Seed for random number generation. Default value is 1.

pop_size::Int64: Size of the population. Default value is 1000.

I_init::Int64: Initial number of infected individuals. Default value is 200.

H_init::Int64: Initial number of hospitalized individuals. Default value is 20.

beta::Float64: Transmission rate. Default value is 0.001.

gamma::Float64: Rate of exposed individuals becoming infectious. Default value is 1/4.

nu::Float64: Rate of infected individuals recovering or getting hospitalized. Default value is 1/7.

epsilon::Float64: Rate of hospitalized individuals recovering. Default value is 1/5.

w_init::Float64: Probability of an infected individual becoming hospitalized. Default value is 0.35.

Returns

df::DataFrame: A DataFrame containing the simulation data with columns Time, S, E, I, H, and R.

UCIWWEIHR.generate_simulation_data_uciwweihr — Function

Generating Simulation Data for UCIWWEIHR ODE Compartmental Based Model

To generate simulation data using the UCIWWEIHR ODE compartmental based model, you can use the generate_simulation_data_uciwweihr function defined in the UCIWWEIHR.jl package. This function allows you to customize various parameters for the simulation.

Function Signature

Arguments

time_points::Int64: Number of time points wanted for simulation. Default value is 150.

seed::Int64: Seed for random number generation. Default value is 1.

E_init::Int64: Initial number of exposed individuals. Default value is 200.

I_init::Int64: Initial number of infected individuals. Default value is 100.

H_init::Int64: Initial number of hospitalized individuals. Default value is 20.

gamma::Float64: Rate of incubation. Default value is 1/4.

nu::Float64: Rate of leaving the infected compartment. Default value is 1/7.

epsilon::Float64: Rate of hospitalization recovery. Default value is 1/5.

rho_gene::Float64: Contribution of infected individual's pathogen genome into wastewater. Default value is 0.011.

tau::Float64: Scale/variation of the log concentration of pathogen genome in wastewater. Default value is 0.1.

df::Float64: Degrees of freedom for generalized t distribution for log concentration of pathogen genome in wastewater. Default value is 29.

sigma_hosp::Float64: Standard deviation for the negative binomial distribution for hospital data. Default value is 800.

Rt_init::Float64: Initial value of the time-varying reproduction number. Default value is 1.

sigma_Rt::Float64: Standard deviation for random walk of time-varying reproduction number. Default value is sqrt(0.02).

w_init::Float64: Initial value of the time-varying hospitalization rate. Default value is 0.35.

sigma_w::Float64: Standard deviation for random walk of time-varying hospitalization rate. Default value is sqrt(0.02).

Returns

df::DataFrame: A DataFrame containing the simulation data with columns obstimes, log_ww_conc, hosp, and rt.

UCIWWEIHR.power — Method

power(a,b)

Raise `a` to the `b` power

UCIWWEIHR.save_plots_to_docs — Method

save_plots_to_docs(plot, filename; format = "png")

Saves plots to docs/plots directory.

Function created by Christian Bernal Zelaya.

UCIWWEIHR.startswith_any — Method

startswith_any(name, patterns)

Checks if the name of time varying paramter starts with any of the patterns.

Function created by Christian Bernal Zelaya.

UCIWWEIHR.uciwweihr_fit — Function

uciwweihr_fit(...)

This is the sampler for the bayesian semi-parametric model for the wastewater EIHR compartmental model. The defaults for this fuction will follow those of the default simulation in generatesimulationdatawweihr.jl function.

Arguments

data_hosp: An array of hospital data.

data_wastewater: An array of pathogen genome concentration in localized wastewater data.

obstimes: An array of timepoints for observed hosp/wastewater.

priors_only::Bool=false: A boolean to indicate if only priors are to be sampled.

n_samples::Int64=500: Number of samples to be drawn.

n_chains::Int64=1: Number of chains to be run.

seed::Int64=2024: Seed for the random number generator.

E_init_sd::Float64=50.0: Standard deviation for the initial number of exposed individuals.

E_init_mean::Int64=200: Mean for the initial number of exposed individuals.

I_init_sd::Float64=20.0: Standard deviation for the initial number of infected individuals.

I_init_mean::Int64=100: Mean for the initial number of infected individuals.

H_init_sd::Float64=5.0: Standard deviation for the initial number of hospitalized individuals.

H_init_mean::Int64=20: Mean for the initial number of hospitalized individuals.

gamma_sd::Float64=0.02: Standard deviation for the rate of incubation.

log_gamma_mean::Float64=log(1/4): Mean for the rate of incubation on log scale.

nu_sd::Float64=0.02: Standard deviation for the rate of leaving the infected compartment.

log_nu_mean::Float64=log(1/7): Mean for the rate of leaving the infected compartment on the log scale.

epsilon_sd::Float64=0.02: Standard deviation for the rate of hospitalization recovery.

log_epsilon_mean::Float64=log(1/5): Mean for the rate of hospitalization recovery on the log scale.

rho_gene_sd::Float64=0.02: Standard deviation for the rho prior.

log_rho_gene_mean::Float64=log(0.011): Mean for the row prior on log scale.

tau_sd::Float64=0.02: Standard deviation for the scale/variation of the log scale data.

log_tau_mean::Float64=log(0.1): Mean for the scale/variation of the log scale data on log scale itself.

df_shape::Float64=2.0: Shape parameter for the gamma distribution.

df_scale::Float64=10.0: Scale parameter for the gamma distribution.

sigma_hosp_sd::Float64=50.0: Standard deviation for the negative binomial distribution for hospital data.

sigma_hosp_mean::Float64=500.0: Mean for the negative binomial distribution for hospital data.

Rt_init_sd::Float64=0.3: Standard deviation for the initial value of the time-varying reproduction number.

Rt_init_mean::Float64=0.2: Mean for the initial value of the time-varying reproduction number.

sigma_Rt_sd::Float64=0.2: Standard deviation for normal prior of log time-varying reproduction number standard deviation.

sigma_Rt_mean::Float64=-3.0: Mean for normal prior of log time-varying reproduction number standard deviation.

w_init_sd::Float64=0.1: Standard deviation for the initial value of the time-varying hospitalization rate.

w_init_mean::Float64=log(0.35): Mean for the initial value of the time-varying hospitalization rate.

sigma_w_sd::Float64=0.2: Standard deviation for normal prior of log time-varying hospitalization rate standard deviation.

sigma_w_mean::Float64=-3.5: Mean for normal prior of time-varying hospitalization rate standard deviation.

param_change_times::Array{Float64}: An array of timepoints where the parameters change.

Returns

Samples from the posterior or prior distribution.

UCIWWEIHR.uciwweihr_model — Function

uciwweihr_model(...)

This is the bayesian semi-parametric model for the wastewater EIHR compartmental model. The defaults for this fuction will follow those of the default simulation in generatesimulationdatawweihr.jl function.

Arguments

data_hosp: An array of hospital data.

data_wastewater: An array of pathogen genome concentration in localized wastewater data.

obstimes: An array of timepoints for observed hosp/wastewater.

param_change_times: An array of timepoints where the parameters change.

E_init_sd::Float64=50.0: Standard deviation for the initial number of exposed individuals.

E_init_mean::Int64=200: Mean for the initial number of exposed individuals.

I_init_sd::Float64=20.0: Standard deviation for the initial number of infected individuals.

I_init_mean::Int64=100: Mean for the initial number of infected individuals.

H_init_sd::Float64=5.0: Standard deviation for the initial number of hospitalized individuals.

H_init_mean::Int64=20: Mean for the initial number of hospitalized individuals.

gamma_sd::Float64=0.02: Standard deviation for the rate of incubation.

log_gamma_mean::Float64=log(1/4): Mean for the rate of incubation on log scale.

nu_sd::Float64=0.02: Standard deviation for the rate of leaving the infected compartment.

log_nu_mean::Float64=log(1/7): Mean for the rate of leaving the infected compartment on the log scale.

epsilon_sd::Float64=0.02: Standard deviation for the rate of hospitalization recovery.

log_epsilon_mean::Float64=log(1/5): Mean for the rate of hospitalization recovery on the log scale.

rho_gene_sd::Float64=0.02: Standard deviation for the rho prior.

log_rho_gene_mean::Float64=log(0.011): Mean for the row prior on log scale.

tau_sd::Float64=0.02: Standard deviation for the scale/variation of the log scale data.

log_tau_mean::Float64=log(0.1): Mean for the scale/variation of the log scale data on log scale itself.

df_shape::Float64=2.0: Shape parameter for the gamma distribution.

df_scale::Float64=10.0: Scale parameter for the gamma distribution.

sigma_hosp_sd::Float64=50.0: Standard deviation for the negative binomial distribution for hospital data.

sigma_hosp_mean::Float64=500.0: Mean for the negative binomial distribution for hospital data.

Rt_init_sd::Float64=0.3: Standard deviation for the initial value of the time-varying reproduction number.

Rt_init_mean::Float64=0.2: Mean for the initial value of the time-varying reproduction number.

sigma_Rt_sd::Float64=0.2: Standard deviation for normal prior of log time-varying reproduction number standard deviation.

sigma_Rt_mean::Float64=-3.0: Mean for normal prior of log time-varying reproduction number standard deviation.

w_init_sd::Float64=0.1: Standard deviation for the initial value of the time-varying hospitalization rate.

w_init_mean::Float64=log(0.35): Mean for the initial value of the time-varying hospitalization rate.

sigma_w_sd::Float64=0.2: Standard deviation for normal prior of log time-varying hospitalization rate standard deviation.

sigma_w_mean::Float64=-3.5: Mean for normal prior of time-varying hospitalization rate standard deviation.

UCIWWEIHR.uciwweihr_visualizer — Method

uciwweihr_visualizer(...)

Default visualizer for results of the UCIWWEIHR model, includes posterior/priors of generated quantities and posterior predictive samples for forecasting. Forecasting plots will have the observed data alongside.

Arguments

pp_sampeles: Posterior predictive samples from the posterior/prior distribution, index 1 in uciwweihrgqpp output.

gq_samples: Generated quantities samples from the posterior/prior distribution, index 2 in uciwweihrgqpp output.

data_hosp: An array of hospital data.

data_wastewater: An array of pathogen genome concentration in localized wastewater data.

obstimes: An array of timepoints for observed hosp/wastewater.

param_change_times: An array of timepoints where the parameters change.

actual_rt_vals: An array of actual Rt values if user has access to them assumed to be on a daily scale. This typically will come from some simulation. Default is nothing.

actual_w_t: An array of actual w_t values if user has access to them assumed to be on a daily scale. This typically will come from some simulation. Default is nothing.

desired_params: A list of lists of parameters to visualize. Each list will be visualized in a separate plot. Default is [["Einit", "Iinit", "Hinit"], ["gamma", "nu", "epsilon"], ["rhogene", "tau", "df"], ["sigma_hosp"]].

time_varying_params: A list of time varying parameters to visualize. Default is ["rtvals", "wt"].

quantiles: A list of quantiles to calculate for ploting uncertainty. Default is [0.5, 0.8, 0.95].

mcmcdaigs::Bool=true: A boolean to indicate if user wants to visualize mcmc diagnosis plots and Effective Sample Size(ESS).

time_varying_plots::Bool=true: A boolean to indicate if user wants to visualize time varying parameters.

save_plots::Bool=false: A boolean to indicate if user wants to save the plots as pngs into a plots folder.