Reference

Contents

Reference

Contents

Index

Index

UCIWWEIHR.uciwweihr_model_params

UCIWWEIHR.uciwweihr_sim_params

UCIWWEIHR.ChainsCustomIndex

UCIWWEIHR.NegativeBinomial2

UCIWWEIHR.calculate_quantiles

UCIWWEIHR.create_uciwweihr_model_params

UCIWWEIHR.create_uciwweihr_sim_params

UCIWWEIHR.eihr_ode!

UCIWWEIHR.generate_colors

UCIWWEIHR.generate_logit_normal_random_walk

UCIWWEIHR.generate_random_walk

UCIWWEIHR.generate_simulation_data_agent

UCIWWEIHR.generate_simulation_data_uciwweihr

UCIWWEIHR.is_time_varying_above_n

UCIWWEIHR.mcmcdiags_vis

UCIWWEIHR.non_time_varying_param_vis

UCIWWEIHR.power

UCIWWEIHR.predictive_param_vis

UCIWWEIHR.repeat_last_n_elements

UCIWWEIHR.repeated_forecast

UCIWWEIHR.save_plots_to_docs

UCIWWEIHR.startswith_any

UCIWWEIHR.uciwweihr_fit

UCIWWEIHR.uciwweihr_model

UCIWWEIHR.uciwweihr_model_params — Type

uciwweihr_model_params

Struct for holding parameters used in the UCIWWEIHR ODE compartmental model. Use create_uciwweihr_model_params to create an instance of this struct.

Fields

E_init_sd::Float64=50.0: Standard deviation for the initial number of exposed individuals.

E_init_mean::Int64=200: Mean for the initial number of exposed individuals.

I_init_sd::Float64=20.0: Standard deviation for the initial number of infected individuals.

I_init_mean::Int64=100: Mean for the initial number of infected individuals.

H_init_sd::Float64=5.0: Standard deviation for the initial number of hospitalized individuals.

H_init_mean::Int64=20: Mean for the initial number of hospitalized individuals.

gamma_sd::Float64=0.02: Standard deviation for the rate of incubation.

log_gamma_mean::Float64=log(1/4): Mean for the rate of incubation on log scale.

nu_sd::Float64=0.02: Standard deviation for the rate of leaving the infected compartment.

log_nu_mean::Float64=log(1/7): Mean for the rate of leaving the infected compartment on the log scale.

epsilon_sd::Float64=0.02: Standard deviation for the rate of hospitalization recovery.

log_epsilon_mean::Float64=log(1/5): Mean for the rate of hospitalization recovery on the log scale.

rho_gene_sd::Float64=0.02: Standard deviation for the rho prior.

log_rho_gene_mean::Float64=log(0.011): Mean for the row prior on log scale.

sigma_wastewater::Float64=log(0.1): Standard deviation for the negative binomial distribution for wastewater data. Not infered.

sigma_hosp::Float64=500.0: Standard deviation for the negative binomial distribution for hospital data. Not infered.

Rt_init_sd::Float64=0.3: Standard deviation for the initial value of the time-varying reproduction number.

Rt_init_mean::Float64=0.2: Mean for the initial value of the time-varying reproduction number.

sigma_Rt_sd::Float64=0.2: Standard deviation for normal prior of log time-varying reproduction number standard deviation.

sigma_Rt_mean::Float64=-3.0: Mean for normal prior of log time-varying reproduction number standard deviation.

w_init_sd::Float64=0.1: Standard deviation for the initial value of the time-varying hospitalization rate.

w_init_mean::Float64=log(0.35): Mean for the initial value of the time-varying hospitalization rate.

sigma_w_sd::Float64=0.2: Standard deviation for normal prior of log time-varying hospitalization rate standard deviation.

sigma_w_mean::Float64=-3.5: Mean for normal prior of time-varying hospitalization rate standard deviation.

UCIWWEIHR.uciwweihr_sim_params — Type

uciwweihr_sim_params

Struct for holding parameters used in the UCIWWEIHR ODE compartmental model simulation. Use create_uciwweihr_sim_params to create an instance of this struct.

Fields

time_points::Int64: Number of time points for the simulation.

seed::Int64: Seed for random number generation.

E_init::Int64: Initial number of exposed individuals.

I_init::Int64: Initial number of infected individuals.

H_init::Int64: Initial number of hospitalized individuals.

gamma::Float64: Rate of incubation.

nu::Float64: Rate of leaving the infected compartment.

epsilon::Float64: Rate of hospitalization recovery.

rho_gene::Float64: Contribution of infected individual's pathogen genome into wastewater.

tau::Float64: Scale/variation of the log concentration of pathogen genome in wastewater.

df::Float64: Degrees of freedom for generalized t-distribution for log concentration of pathogen genome in wastewater.

sigma_hosp::Float64: Standard deviation for the negative binomial distribution for hospital data.

Rt::Union{Float64, Vector{Float64}}: Initial value or time series of the time-varying reproduction number.

sigma_Rt::Float64: Standard deviation for random walk of time-varying reproduction number.

w::Union{Float64, Vector{Float64}}: Initial value or time series of the time-varying hospitalization rate.

sigma_w::Float64: Standard deviation for random walk of time-varying hospitalization rate.

rt_init::Float64: Initial value of the time-varying reproduction number, NOT USER SPECIFIED create_uciwweihr_params TAKES CARE OF THIS.

w_init::Float64: Initial value of the time-varying hospitalization rate, NOT USER SPECIFIED create_uciwweihr_params TAKES CARE OF THIS.

UCIWWEIHR.ChainsCustomIndex — Method

ChainsCustomIndex(c::Chains, indices_to_keep::BitMatrix)

Reduce Chains object to only wanted indices.

Function created by Damon Bayer.

UCIWWEIHR.NegativeBinomial2 — Method

Create a re-parametrized negative binomial distribution in terms of mean and overdispersion.

Arguments

μ: Mean of the distribution.

ϕ: Overdispersion parameter.

Returns

A Distributions.NegativeBinomial distribution object.

UCIWWEIHR.calculate_quantiles — Method

calculate_quantiles(df, chain, var_prefix, quantiles)

Calculate quantiles for a given chain and variable prefix. Quantiles can be any user desired quantile.

UCIWWEIHR.create_uciwweihr_model_params — Method

create_uciwweihr_model_params(; kwargs...)

Creates a uciwweihr_sim_params struct with the option to either use a predetermined Rt and w or generate them as random walks.

Arguments

kwargs...: Named arguments corresponding to the fields in uciwweihr_sim_params.

Returns

params::uciwweihr_sim_params: A struct with simulation parameters.

UCIWWEIHR.create_uciwweihr_sim_params — Method

createuciwweihrsim_params(; kwargs...)

Creates a uciwweihr_sim_params struct with the option to either use a predetermined Rt and w or generate them as random walks.

Arguments

kwargs...: Named arguments corresponding to the fields in uciwweihr_sim_params.

Returns

params::uciwweihr_sim_params: A struct with simulation parameters.

UCIWWEIHR.eihr_ode! — Method

eihr_ode!(du, u, p, t)

Calculate the ordinary differential equations (ODEs) for the EIHR model.

Parameters:

du: Array{Float64,1} - The derivative of the state variables.

u: Array{Float64,1} - The current state variables.

p: Tuple{Float64,Float64,Float64,Float64,Float64} - The model parameters (alpha, gamma, nu, w, epsilon).

t: Float64 - The current time.

Returns:

du: Array{Float64,1} - The derivative of the state variables.

UCIWWEIHR.generate_colors — Method

generate_ribbon_colors(number_of_colors)

Generates a vector with colors for ribbons in plots.

UCIWWEIHR.generate_logit_normal_random_walk — Method

generate_logit_normal_random_walk(time_points::Int64, sigma::Float64, init_val::Float64)

Generates a logit-normal random walk time series.

Arguments

time_points::Int64: Number of time points.

sigma::Float64: Standard deviation of the random walk in logit space.

init_val::Float64: Initial value of the random walk on the probability scale.

Returns

walk::Vector{Float64}: Generated random walk on the probability scale.

UCIWWEIHR.generate_random_walk — Method

generate_random_walk(time_points::Int64, sigma::Float64, init_val::Float64)

Generates a random walk time series.

Arguments

time_points::Int64: Number of time points.

sigma::Float64: Standard deviation of the random walk.

init_val::Float64: Initial value of the random walk.

Returns

walk::Vector{Float64}: Generated random walk.

UCIWWEIHR.generate_simulation_data_agent — Function

Generating Simulation Data for Agent Based Model

To generate simulation data using the agent based model, you can use the generate_simulation_data_agent function defined in the UCIWWEIHR.jl package. This function allows you to customize various parameters for the simulation. NOT FINISHED, STILL NEEDS WW AND RT

Function Signature

Arguments

seed::Int64: Seed for random number generation. Default value is 1.

pop_size::Int64: Size of the population. Default value is 1000.

I_init::Int64: Initial number of infected individuals. Default value is 200.

H_init::Int64: Initial number of hospitalized individuals. Default value is 20.

beta::Float64: Transmission rate. Default value is 0.001.

gamma::Float64: Rate of exposed individuals becoming infectious. Default value is 1/4.

nu::Float64: Rate of infected individuals recovering or getting hospitalized. Default value is 1/7.

epsilon::Float64: Rate of hospitalized individuals recovering. Default value is 1/5.

w_init::Float64: Probability of an infected individual becoming hospitalized. Default value is 0.35.

Returns

df::DataFrame: A DataFrame containing the simulation data with columns Time, S, E, I, H, and R.

UCIWWEIHR.generate_simulation_data_uciwweihr — Method

generate_simulation_data(params::UCIWWEIHRParams)

Generates simulation data for the UCIWWEIHR ODE compartmental model.

Arguments

params::uciwweihr_sim_params: Struct containing parameters for the simulation.

Returns

df::DataFrame: A DataFrame containing the simulation data with columns obstimes, log_ww_conc, hosp, rt, and wt.

UCIWWEIHR.is_time_varying_above_n — Method

is_time_varying_above_n(name, n)

Checks if the time varying parameter is above a given time point.

UCIWWEIHR.mcmcdiags_vis — Method

mcmcdiags_vis(...)

Default visualizer for results of the UCIWWEIHR model, includes posterior/priors of generated quantities and posterior predictive samples for forecasting. Forecasting plots will have the observed data alongside.

Arguments

gq_samples: Generated quantities samples from the posterior/prior distribution, index 2 in uciwweihrgqpp output.

desired_params: A list of lists of parameters to visualize. Each list will be visualized in a separate plot. Default is [["Einit", "Iinit", "Hinit"], ["gamma", "nu", "epsilon"], ["rhogene", "tau", "df"], ["sigma_hosp"]].

actual_non_time_varying_vals::uciwweihr_sim_params: A uciwweihrsimparams object of actual non-time varying parameter values if user has access to them. Default is nothing.

save_plots::Bool=false: A boolean to indicate if user wants to save the plots as pngs into a plots folder.

plot_name_to_save: A string to indicate the name of the plot to save. Default is "mcmcdiagnosisplots".

UCIWWEIHR.non_time_varying_param_vis — Method

non_time_varying_param_vis(...)

Used in the uciwweihr_visualizer to create visuals for non-time varying parameters.

Arguments

build_params::uciwweihr_model_params: A struct of model parameters used to build gq_samples, used only if user desired priors next to posteriors.

data_hosp: Hospitalization data, used only if user desired priors next to posteriors.

data_wastewater: Wastewater data, if model does not use this do not specify this, if user desires priors next to plot (do not specify if you do not want prior plots).

obstimes_hosp: An array of time points for hospital data, used only if user desired priors next to posteriors.

obstimes_wastewater: An array of time points for wastewater data, used only if user desired priors next to posteriors.

param_change_times: An array of time points where the parameters change, used only if user desired priors next to posteriors.

seed: An integer to set the seed for reproducibility, used only if user desired priors next to posteriors.

forecast: A boolean to indicate if user wants to forecast, used only if user desired priors next to posteriors.

forecast_weeks: An integer to indicate the number of weeks to forecast, used only if user desired priors next to posteriors.

gq_samples: Generated quantities samples from the posterior/prior distribution, index 2 in uciwweihrgqpp output.

desired_params: A list of lists of parameters to visualize. Each list will be visualized in a separate plot. Default is any parameter not in this list : ["alphat", "wt", "rtvals", "loggenes_mean", "H"]

actual_non_time_varying_vals::uciwweihr_sim_params: A uciwweihrsimparams object of actual non-time varying parameter values if user has access to them. Default is nothing.

save_plots::Bool=false: A boolean to indicate if user wants to save the plots as pngs into a plots folder.

plot_name_to_save: A string to indicate the name of the plot to save. Default is "mcmcnontimevaryingparameterplots".

UCIWWEIHR.power — Method

power(a,b)

Raise `a` to the `b` power

UCIWWEIHR.predictive_param_vis — Method

predictive_param_vis(...)

Used in the uciwweihr_visualizer to create visuals for wastewater data and hospitalization data.

Arguments

pp_samples: A DataFrame of posterior or prior predictive samples.

data_wastewater: An array of actual wastewater values if user has access to them assumed, using time scale of observed time points. Default is nothing.

data_hosp: An array of actual hospitalization values if user has access to them assumed, , using time scale of observed time points. Default is nothing.

forecast_weeks: An integer of the number of weeks forecasted. Default is 0.

vars_to_pred: A list of variables to predict. Default is ["datawastewater", "datahosp"].

quantiles: A list of quantiles to calculate for ploting uncertainty. Default is [0.5, 0.8, 0.95].

bayes_dist_type: A string to indicate if user is using Posterior or Prior distribution. Default is "Posterior".

save_plots::Bool=false: A boolean to indicate if user wants to save the plots as pngs into a plots folder.

plot_name_to_save: A string to indicate the name of the plot to save. Default is "mcmcpredparameter_plots".

UCIWWEIHR.repeat_last_n_elements — Method

repeat_last_n_elements(x::Vector{T}, n::Int, w::Int) where T

Modifies a given array so that the last n elements are repeated w times.

UCIWWEIHR.repeated_forecast — Method

repeated_forecast(...)

This is the function to make repreated forecast for a given forecast time span, n_forecast_weeks, and for given time points, forecast_points. Plots can be made for these forecasts. The output is an array of uciwweihr_gq_pp results for each forecast_points.

Arguments

data_hosp: The hospitalization data.

data_wastewater: The wastewater data.

obstimes_hosp: The time points for the hospitalization data.

obstimes_wastewater: The time points for the wastewater data.

n_samples: The number of samples to draw from the posterior.

param_change_times: The time points where the parameters change.

params::uciwweihr_model_params: The model parameters.

n_forecast_weeks: The number of weeks to forecast.

forecast_points: The time points to forecast, thees points should be present in obstimes_hosp.

Returns

An array of uciwweihr_gq_pp resuts and timeseries used for building for each forecast_points.

UCIWWEIHR.save_plots_to_docs — Method

save_plots_to_docs(plot, filename; format = "png")

Saves plots to docs/plots directory.

UCIWWEIHR.startswith_any — Method

startswith_any(name, patterns)

Checks if the name of time varying paramter starts with any of the patterns.

UCIWWEIHR.uciwweihr_fit — Method

uciwweihr_fit(...)

This is the sampler for the bayesian semi-parametric model for the wastewater EIHR compartmental model. The defaults for this fuction will follow those of the default simulation in generatesimulationdatawweihr.jl function.

Arguments

data_hosp: An array of hospital data.

data_wastewater: An array of pathogen genome concentration in localized wastewater data. If this is not avaliable, the model used will be one that only uses hospital data.

obstimes: An array of timepoints for observed hosp/wastewater.

priors_only::Bool=false: A boolean to indicate if only priors are to be sampled.

n_samples::Int64=500: Number of samples to be drawn.

n_chains::Int64=1: Number of chains to be run.

seed::Int64=2024: Seed for the random number generator.

params::uciwweihr_model_params: A struct containing parameters for the model.

Returns

Samples from the posterior or prior distribution.

UCIWWEIHR.uciwweihr_model — Method

uciwweihr_model(...)

This is the bayesian semi-parametric model for the wastewater EIHR compartmental model. The defaults for this fuction will follow those of the default simulation in generatesimulationdatawweihr.jl function.

Arguments

data_hosp: An array of hospital data.

data_wastewater: An array of pathogen genome concentration in localized wastewater data. If this is not avaliable, the model used will be one that only uses hospital data.

obstimes_hosp: An array of timepoints for observed hospital data.

obstimes_wastewater: An array of timepoints for observed wastewater data.

param_change_times: An array of timepoints where the parameters change.

params::uciwweihr_model_params: A struct containing parameters for the model.